+++ NEWS.md
@@ -1,3 +1,10 @@
+# contactdata 0.2
+
+* Addition of a new function `age_df_countries()` to get population
+by age for all the countries for which we provide contact data.
+This is motivated by the fact that models using contact data by
+age will almost surely need population data by age as well.
+
 # contactdata 0.1
 
 * First release